Move logic from use_submit.tsx to advanced_text_editor.tsx. Create eliminateMentionNicknameOrFullName, and this function will be reused.
